requirements.txt
djangorestframework==3.2
jupyter==1.0.0
ipython==4.0.0
psycopg2==2.6.1
django-filter==0.11.0
-e git+https://github.com/django-extensions/django-extensions.git@92008a391c6357d7896a65ea74ac103664e21ddc#egg=django_extensions-master

